CSS STYLE AND LAYOUT AND SOME SMALL JAVASCRIPT CHANGES
All html elements should now sit in the correct positions on the page
and should have consistent styling. A couple of divs where added to the
html to make it easier to position elements that rellevant to each
other and some extra classes added.

Additions were made to the javascript to let elements be visible or not
at the appropriate times.

Two if functions were added to players to prevent the player names from being
empty strings.